requirements.txt
discord.py
requests
Pillow